refactor: move memory to a dedicated submodule with internal keyword
