+++ src/libs/satellite/src/memory/mod.rs
@@ -1,0 +1,2 @@
+pub mod internal;
+mod manager;
