+++ src/libs/satellite/src/types.rs
@@ -1,7 +1,7 @@
 pub mod state {
     use crate::auth::types::state::AuthenticationHeapState;
     use crate::db::types::state::{DbHeapState, DbRuntimeState, DbStable};
-    use crate::memory::init_stable_state;
+    use crate::memory::internal::init_stable_state;
     use crate::storage::types::state::{AssetsStable, ContentChunksStable};
     use candid::CandidType;
     use junobuild_shared::types::state::Controllers;
